+++ src/hooks/EditorTimerPause.hpp
@@ -1,0 +1,12 @@
+#include <Geode/modify/EditorPauseLayer.hpp>
+#include "../timer/Layers/TimerSettingsLayer.hpp"
+#include "../timer/Layers/TimerSettingsButton.hpp"
+#include "EditorUITimer.hpp"
+
+using namespace geode::prelude;
+
+class $modify(EditorTimerPause, EditorPauseLayer) {
+	bool init(LevelEditorLayer* lvl);
+	void onResume(CCObject* sender);
+	// void onTimerSettings(CCObject* sender);
+};