+++ src/hooks/TimerPauseLayer.hpp
@@ -1,0 +1,10 @@
+#include "../timer/Layers/TimerSettingsLayer.hpp"
+#include "../timer/Layers/TimerSettingsButton.hpp"
+#include "../timer/Layers/TimerSettingsButton.hpp"
+#include <Geode/modify/PauseLayer.hpp>
+
+using namespace geode::prelude;
+
+class $modify(TimerPauseLayer, PauseLayer) {
+	void customSetup();
+};