+++ src/timer/Layers/TimerSettingsLayer.hpp
@@ -1,0 +1,33 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+#include <Geode/ui/GeodeUI.hpp>
+#include <Geode/ui/Popup.hpp>
+#include "../../hooks/EditorUITimer.hpp"
+#include "../../hooks/TimerPlayLayer.hpp"
+
+using namespace geode::prelude;
+
+class TimerSettingsLayer : public Popup<CCNode* const&> {
+
+// friend class modSettingsEvent;
+
+public:
+    static TimerSettingsLayer* create(CCNode* const& menuID);
+    static void disableButton(CCNode* node, bool enable);
+    static TimerSettingsLayer* get();
+
+protected:
+
+    CCNode* m_menuID; // ptr to buttons attatched to this layer
+    EventListener<SettingChangedFilter>* m_listener;
+    bool paused;
+    CCMenu* menuPointer;
+  
+    bool setup(CCNode* const& menuID);
+    void resetTimer(CCObject* sender);
+    void changeSettings(CCObject* sender);
+    void pauseTime(CCObject* sender);
+    void updateButtons();
+    ~TimerSettingsLayer();
+};