+++ src/timer/TimerLayer.hpp
@@ -2,18 +2,16 @@
 #include <Geode/Geode.hpp>
 using namespace geode::prelude;
 
-// #include "TimerEvent.hpp"
-
 using countTask = Task<bool, int>;
 
 class TimerLayer : public Popup<std::string const&> {
-    public:
-        static TimerLayer* create(std::string const& waitTime);
+public:
+    static TimerLayer* create(std::string const& waitTime);
 
-    protected:
-        void countDown(countTask::Event* event);
-        bool setup(std::string const& waitTime);
-        void onClick(CCObject* sender);
-        virtual void onClose(cocos2d::CCObject*);
+protected:
+    void countDown(countTask::Event* event);
+    bool setup(std::string const& waitTime);
+    void onClick(CCObject* sender);
+    virtual void onClose(cocos2d::CCObject*);
 
 };