+++ changelog.md
@@ -1,2 +1,11 @@
+<!-- # 1.1.0
+- Added timer modifications layer
+    - Reset Timer
+    - Pause Timer
+    - Change Settings
+- Fixed bugs
+- Optimized -->
+# 1.1.0-alpha1
+- inital testing release
 # 1.0.0
-- Edit this file to change your mod's changelog.
+- Implemented Timer